requirements.txt
selenium
pandas
google-generativeai
matplotlib
seaborn
wordcloud
tensorflow
transformers
pyyaml
python-dotenv
streamlit
plotly
Jinja2
spacy
emoji
networkx
jinja2
subprocess
base64
glob
json